Carlos

+++ appSpring/src/main/java/appSpring/restController/ActionRestController.java
@@ -1,5 +1,7 @@
 package appSpring.restController;
 
+import java.util.Calendar;
+import java.util.Date;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,16 +11,18 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.fasterxml.jackson.annotation.JsonView;
 
 import appSpring.model.Action;
+import appSpring.model.Fine;
 import appSpring.model.Resource;
 import appSpring.model.ResourceCopy;
 import appSpring.model.User;
 import appSpring.repository.ActionRepository;
+import appSpring.service.ResourceService;
+import appSpring.service.UserService;
 
 @RestController
 @RequestMapping("/api/loan")
@@ -28,14 +32,38 @@ public class ActionRestController {
 
 	@Autowired
 	private ActionRepository actionRepository;
+	@Autowired
+	private UserService userService;
+	@Autowired
+	private ResourceService resourceService;
 
 	@RequestMapping(value = "/", method = RequestMethod.POST)
-	@ResponseStatus(HttpStatus.CREATED)
-	public Action postAction(@RequestBody Action loan) {
+	public ResponseEntity<Action> postAction(@RequestBody Action loan) {
 
-		actionRepository.save(loan);
-
-		return loan;
+		Calendar today = Calendar.getInstance();
+		today.set(Calendar.HOUR_OF_DAY, 0);
+		User user = loan.getUser();
+		List<Fine> fines = user.getPenalties();
+		for (Fine fine : fines) {
+			Date currentDate = new Date();
+			if (currentDate.before(fine.getFinishDate()) || (user.getisBanned()) || (user.getAvaibleLoans() == 0)) {
+				return new ResponseEntity<>(HttpStatus.CONFLICT);
+			}
+		}
+		if (loan.getResource().getResource().getNoReservedCopies().contains(loan.getResource().getLocationCode())) {
+			Resource resource = loan.getResource().getResource();
+			resource.getNoReservedCopies().remove(loan.getResource().getLocationCode());
+			user.setAvaibleLoans(user.getAvaibleLoans()-1);
+			if (resource.getNoReservedCopies().isEmpty()) {
+				resource.setAvaibleReserve(!resource.getAvaibleReserve());
+			}
+			userService.save(user);
+			resourceService.save(resource);
+			actionRepository.save(loan);
+			return new ResponseEntity<>(loan, HttpStatus.CREATED);
+		} else {
+			return new ResponseEntity<>(HttpStatus.CONFLICT);
+		}
 	}
 
 	@JsonView(LoanDetail.class)
@@ -69,6 +97,9 @@ public class ActionRestController {
 
 		Action loan = actionRepository.findOne(id);
 		if (loan != null) {
+			if (loan.getDateLoanReturn() == null) {
+				return new ResponseEntity<>(HttpStatus.CONFLICT);
+			}
 			actionRepository.delete(loan);
 			return new ResponseEntity<>(loan, HttpStatus.OK);
 		} else {
